cmd: Move bccCmd() in a dedicated package and export it.
